feat: 게시글 삭제 시, 댓글 삭제 기능 구현 - Replies 테이블 수정 	- DELETED 필드를 추가하여 soft delete로 동작하도록 수정
- 게시글 삭제 시, 댓글도 같이 삭제될 수 있게 구현

+++ src/main/java/codesquad/springcafe/articles/service/ArticleService.java
@@ -5,6 +5,7 @@ import codesquad.springcafe.articles.model.dto.ArticleUpdateDto;
 import codesquad.springcafe.articles.model.dto.ReplyCreationRequest;
 import codesquad.springcafe.articles.model.dto.ReplyViewDto;
 import codesquad.springcafe.articles.repository.ArticleRepository;
+import codesquad.springcafe.exception.ArticleAccessException;
 import codesquad.springcafe.exception.ArticleNotFoundException;
 import codesquad.springcafe.articles.model.Article;
 import codesquad.springcafe.articles.model.dto.ArticleCreationRequest;
@@ -53,6 +54,12 @@ public class ArticleService {
     }
 
     public void deleteArticle(long articleId) {
+        Article article = articleRepository.findArticleById(articleId).orElseThrow(() -> new ArticleNotFoundException("게시글을 찾을 수 없습니다."));
+
+        ArrayList<Reply> replies = articleRepository.getReplies(articleId).orElseThrow(() -> new ReplyNotFoundException("댓글을 찾을 수 업습니다."));
+
+        validateArticleAccess(replies, article.getUserId());
+
         articleRepository.deleteArticle(articleId);
     }
 
@@ -79,5 +86,19 @@ public class ArticleService {
     public void deleteReply(long replyId) {
         articleRepository.deleteReply(replyId);
     }
+
+    private void validateArticleAccess(ArrayList<Reply> replies, String userId) {
+
+        // * 댓글이 없는 경우
+        if (replies.isEmpty()) {
+            return;
+        }
+
+        // * 댓글이 있지만, 모든 게시글의 저자가 자신인 경우
+        if (replies.stream().anyMatch(reply -> !reply.getUserId().equals(userId))) {
+            throw new ArticleAccessException("게시글을 삭제할 수 없습니다. [에러 내용] : 작성자가 다른 댓글이 있는 경우 삭제가 불가능합니다.");
+        }
+    }
+
 
 }
